requirements.txt
aioredis==2.0.1
aiosmtplib==1.1.6
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
autopep8==1.6.0
bcrypt==3.2.2
blinker==1.4
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
Deprecated==1.2.13
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
fakeredis==1.8
fastapi==0.75.2
fastapi-mail==1.0.8
greenlet==1.1.2
h11==0.12.0
httpcore==0.14.7
httptools==0.4.0
httpx==0.22.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.16.0
orjson==3.6.8
packaging==21.3
passlib==1.7.4
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.1
PyMySQL==1.0.2
pyparsing==3.0.9
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
redis==4.3.1
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.37
starlette==0.17.1
toml==0.10.2
typing_extensions==4.2.0
ujson==5.3.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
Werkzeug==2.1.2
wrapt==1.14.1
